+++ CMakeLists.txt
@@ -79,13 +79,14 @@ message("BUILD_WITH_ALL: ${BUILD_WITH_ALL}")
 
 if (NINTENDO_3DS)
     set(BUILD_SDL OFF)
+    set(PREFER_SYSTEM_LIBRARIES ON)
 endif()
 
 if (BAREMETALPI)
     set(BUILD_SDL OFF)
 endif()
 
-if(UNIX AND NOT APPLE AND NOT EMSCRIPTEN AND NOT ANDROID AND NOT NINTENDO_3DS)
+if(UNIX AND NOT APPLE AND NOT EMSCRIPTEN AND NOT ANDROID)
     set(LINUX TRUE)
 
     if(CMAKE_HOST_SYSTEM_NAME STREQUAL "FreeBSD")
